+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [44.3.0] 
+- Resources was updated from DIPS.Mobile.DesignTokens
+
 ## [44.2.0] 
 - Resources was updated from DIPS.Mobile.DesignTokens